corrected mispriming bug
Tracked down (after many hours) the very difficult to trace bug that was causing invalid PCR products to be produced.

This had to do with the very convoluted MultiPCRProductAlignmentGroup class and the fact the the groupings['template'] does not actually represent where the primers are being designed in primer3. Instead, this is found in the method `get_template`, which uses the intersection between two regions (slowing down the computation as well). For this reason, `groupings` probably show never be accessed directly.